Move chore files
