+++ src/contexts/RouterContext.tsx
@@ -1,6 +1,6 @@
 import { App } from '../App'
-import { ChoreEdit } from '../views/ChoreEdit/ChoreEdit'
-import { ChoresOverview } from '../views/ChoresOverview'
+import { ChoreEdit } from '../views/Chores/ChoreEdit'
+import { ChoresOverview } from '../views/Chores/ChoresOverview'
 import { Error } from '../views/Error'
 import { Settings } from '../views/Settings/Settings'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
@@ -8,7 +8,7 @@ import { ForgotPasswordView } from '../views/Authorization/ForgotPasswordView'
 import { LoginView } from '../views/Authorization/LoginView'
 import { SignupView } from '../views/Authorization/Signup'
 import { UpdatePasswordView } from '../views/Authorization/UpdatePasswordView'
-import { ChoreView } from '../views/ChoreEdit/ChoreView'
+import { ChoreView } from '../views/Chores/ChoreView'
 import { MyChores } from '../views/Chores/MyChores'
 import { ChoreHistory } from '../views/History/ChoreHistory'
 import { LabelView } from '../views/Labels/LabelView'
